Write fonctionne. Reste Routine et Piezo

+++ tp/tp9/lib_dir/UART.h
@@ -1,12 +1,13 @@
 #ifndef UART_H
 #define UART_H
+#include "includes.h"
 
 class UART {
     public:
         UART();     
         void initialisationUART();
-        void transmissionUART(unsigned int data);
-        unsigned int receptionUART();
+        void transmissionUART(uint8_t data);
+        uint8_t receptionUART();
 };
 
 #endif
